+++ wemake_python_styleguide/presets/types/tree.py
@@ -20,6 +20,7 @@ from wemake_python_styleguide.visitors.ast import (
     naming,
     operators,
     statements,
+    subscripts,
 )
 
 #: Used to store all general visitors to be later passed to checker:
@@ -96,4 +97,7 @@ PRESET: Final = (
 
     # Complexity:
     *complexity.PRESET,
+
+    # Subscripts:
+    subscripts.SubscriptVisitor,
 )