+++ wemake_python_styleguide/violations/consistency.py
@@ -74,6 +74,7 @@ Summary
    OperationSignNegationViolation
    VagueImportViolation
    AdditionAssignmentOnListViolation
+   RedundantSubscriptViolation
 
 Consistency checks
 ------------------
@@ -127,6 +128,7 @@ Consistency checks
 .. autoclass:: OperationSignNegationViolation
 .. autoclass:: VagueImportViolation
 .. autoclass:: AdditionAssignmentOnListViolation
+.. autoclass:: RedundantSubscriptViolation
 
 """
 
@@ -1788,7 +1790,6 @@ class VagueImportViolation(ASTViolation):
     code = 347
 
 
-@final
 class AdditionAssignmentOnListViolation(ASTViolation):
     """
     Forbids usage of += with list arguments.
@@ -1811,3 +1812,29 @@ class AdditionAssignmentOnListViolation(ASTViolation):
 
     error_template = 'Found addition assignment with list argument'
     code = 348
+
+
+@final
+class RedundantSubscriptViolation(ASTViolation):
+    """
+    Forbids the use of redundant components in a subscript's slice.
+
+    Reasoning:
+        We do it for consistency reasons.
+
+    Example::
+
+        # Correct:
+        array[:7]
+        array[3:]
+
+        # Wrong:
+        x[0:7]
+        x[3:None]
+
+    .. versionadded:: 0.13.0
+
+    """
+
+    error_template = 'Found redundant subscript slice: {0}'
+    code = 349