+++ CHANGELOG.md
@@ -19,6 +19,7 @@ We used to have incremental versioning before `0.1.0`.
 - Make ``OveruseOfNoqaCommentViolation`` configurable (`--max-noqa-comments`)
 - Forbid incorrectly swapped variables
 - Forbids to use `+=` with list arguments
+- Forbids to use redundant subscripts (e.g., `[0:7]` or `[3:None]`)
 
 ### Bugfixes